+++ ParProgrammeringLeviAndViktor/InventoryItem.cs
@@ -10,9 +10,9 @@ namespace ParProgrammeringLeviAndViktor
     {
         public string Name { get; private set; }
         public int Quantity { get; private set; }
-        public double Price { get; private set; }
+        public decimal Price { get; private set; }
 
-        protected InventoryItem(string name, int quantity, double price)
+        protected InventoryItem(string name, int quantity, decimal price)
         {
             Name = name;
             Quantity = quantity;
